Splited ContainerReader into seperate classes | code optimisations
