+++ Ps4EditLib/Ps4EditLib.csproj
@@ -49,12 +49,15 @@
     <Compile Include="Exceptions\InvalidHashException.cs" />
     <Compile Include="Extensions\ByteExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="PsRegistry\PsRegistry.cs" />
+    <Compile Include="PsRegistry\Preferences.cs" />
     <Compile Include="PsRegistry\RegFile.cs" />
     <Compile Include="PsRegistry\RegInfo.cs" />
-    <Compile Include="Reader\EntityReader.cs" />
+    <Compile Include="Reader\DataContainerReader.cs" />
+    <Compile Include="Reader\ContainerReaderBase.cs" />
     <Compile Include="Reader\Entry.cs" />
     <Compile Include="Reader\EntryType.cs" />
+    <Compile Include="Reader\IEntityReaderBase.cs" />
+    <Compile Include="Reader\ObfuscatedContainerReader.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
