+++ Ps4EditLib/Reader/IEntityReaderBase.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Ps4EditLib.Reader
+{
+    public interface IEntityReader
+    {
+        List<Entry> Entries { get; }
+        bool ObfuscatedContainer { get; }
+    }
+}
